chore: move everything into module for ResourcesController
